Separado os códigos e primeira versão da visualização grafica
* Criado algumas pastas e adaptado o Makefile
* Criado um menu.h com os métodos do menu
* Primeira versão da visualização grafica
