+++ .gitignore
@@ -1,1 +1,5 @@
-./p
+main
+
+*.png
+*.dot
+obj/*.o
