+++ dados
@@ -1,1 +1,1 @@
-dadosDemograficos/d|faixaEtaria/f|*localExamePreNatal/l|pais/p|regiao/r|estado/e|cidade/c|hospital/h|*dadosEstatisticosPreNatal/dadosE|nroExames/n|*hospitalNascimentoBebe/hospitalN|pais/pais|regiao/r|estado/e|cidade/c|hospital/h|*data/|ano/|semestre/|trimestre/|mes/|dia/|*#produto/p|marca/m|produto/produto|*filial/|pais/|regiao/|estado/|cidade/|filial/|*data/|semestre/|trimestre/|mes/|dia/|*promocao/|tipoPromocao/|promocao/|promocao/|2/|*#
+produto/p|marca/m|produto/produto|*filial/|estado/|cidade/|filial/|*#
