+++ include/menu.h
@@ -1,0 +1,11 @@
+#ifndef MENU_H_
+#define MENU_H_
+
+void menu_principal();
+void menu_continuar();
+void menu_visualizar_grafo();
+void menu_novo_grafo();
+void menu_visualizacao_grafica();
+void menu_sair();
+
+#endif
