+++ src/main.c
@@ -1,8 +1,10 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
-#include <unistd.h>
 
 #include "structs.h"
+#include "menu.h"
+
 #define tamNome 50
 #define tamSigla 10
 
@@ -24,7 +26,6 @@ void imprimiHierarquias (Grafo *g) {
 
 
 void novoGrafo (Grafo* g) {
-
 	printf("\n**********************************************************************");
 	printf("\n(IMPORTANTE)A ordem em que forem inseridas as informações é a ordem da hierarquia\n");
 	printf("Digite quantas dimensões terá o grafo: ");
@@ -204,13 +205,6 @@ void leDados(Grafo *g, int rrn) {
 	//printf("%s",a);
 }
 
-void menu_principal();
-void menu_continuar();
-void menu_visualizar_grafo();
-void menu_novo_grafo();
-void menu_visualizacao_grafica();
-void menu_sair();
-
 int main() {
 /*
 		const char src[50] = "asd";
@@ -224,105 +218,4 @@ int main() {
 */
 	menu_principal();
 	return 0;
-}
-
-/**
- * @brief Método para mostrar o menu principal
- */
-void menu_principal() {
-	int menu_opcao;
-
-	printf("*************************************************\n");
-	printf("** Bem vindo ao Gerador de Grafos de Derivacao **\n");
-	printf("*************************************************\n\n");
-	printf("Menu:\n");
-	printf("[1] Visualizar grafo\n[2] Novo grafo\n[3] Visualizar graficamente um grafo\n[4] Sair\n\n");
-	printf("Opcao: ");
-
-	while (1) {
-		fflush(stdin);
-		scanf("%d", &menu_opcao);
-		if (menu_opcao == 1) {
-			menu_visualizar_grafo();
-		}
-		else if (menu_opcao == 2) {
-			menu_novo_grafo();
-		}
-		else if (menu_opcao == 3) {
-			menu_visualizacao_grafica();
-		}
-		else if (menu_opcao == 4) {
-			menu_sair();
-		}
-	}
-}
-
-/**
- * @brief Método para voltar a menu principal ao
- */
-void menu_continuar() {
-	printf("Aperte [ENTER] para continuar\n");
-	while( getchar() != '\n' );
-	getchar();
-	menu_principal();
-}
-
-/**
- * @brief Método para tratar visualização de grafo
- */
-void menu_visualizar_grafo() {
-	int id_grafo;
-	Grafo leitura;
-
-	printf ("\nDigite o número id do grafo: ");
-	scanf("%d", &id_grafo);
-	
-	createGrafo(&leitura);
-	leDados(&leitura, id_grafo);
-	if (leitura.numHier == 0) {
-		printf("\nNão existe esse grafo");
-	}
-	else {
-		imprimiHierarquias(&leitura);
-	}
-
-	menu_continuar();
-}
-
-/**
- * @brief Método para adicionar um grafo
- */
-void menu_novo_grafo() {
-	Grafo g;
-	createGrafo(&g);
-
-	novoGrafo(&g);
-
-	armazenaDados(&g);
-
-	menu_continuar();
-}
-
-/**
- * @brief Método para visualizar graficamente o grafo
- */
-void menu_visualizacao_grafica() {
-	printf("Opcao 3\n");
-
-	menu_continuar();
-}
-
-/**
- * @brief Método para sair do programa
- */
-void menu_sair() {
-	int i;
-	printf("\nSaindo do programa");
-	for(i = 0;i < 2;i++) {
-    printf(".");
-    fflush(stdout);
-    sleep(1);
-  }
-  printf("\n");
-	exit(1);
 }
